// src/community/ui/data/community/models/GroupMember.ts
export interface GroupMember {
  admin: boolean;
  communityId: string;
  companyId: string;
  companyName: string;
  createdTime: number;
  creatorId: string;
  email: string;
  name: string;
  teamId: string;
  teamName: string;
  nickname: string;
  modifierId: string;
  modifiedTime: number;

  memberId: string;
  groupId: string;
}